flask-app/requirements.txt
flask
flask_login
transformers
torch
json
sqlalchemy.sql